requirements.txt
zeep>=4.2.1
lxml>=4.9.3
xmljson>=0.2.1
python-dotenv>=1.0.0
streamlit>=1.32.0 